Fix replay command in MacOS (#133)
* Fix replay command in MacOS

The replay command wasn't working on MacOS because the HostIO symbols
were striped from cargo-stylus. To fix this issue, we have to disable
stripping in the release profile.

This fixes introduces the following warning:

    warning: profiles for the non root package will be ignored

We have to edit the profile in the non-root package because the root
package is ignored during cargo install. Otherwise, the solution won't
work after the package is published to crates.io. More information about
this issue is available here:

* https://github.com/rust-lang/cargo/issues/8264
* https://github.com/rust-lang/cargo/issues/11599

* Fix missing symbols in MacOS

Add #[used] static variables pointing to the exported functions so the
linked doesn't remove them from the final binary.

* Revert "Fix replay command in MacOS"

This reverts commit b48939457f5a8d1d983c814f66a134b1ff5930ab.